requirements.txt
sentence_transformers
Pillow
transformers
openvino-dev